% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CertaintyEquivalents}
\alias{CertaintyEquivalents}
\title{Transfer error matrices on certainty equivalents for binary lotteries}
\format{
An object of class \code{list} of length 2.
}
\usage{
CertaintyEquivalents
}
\description{
\code{CertaintyEquivalents$unnormalized} and \code{CertaintyEquivalents$normalized}
record the unnormalized and normalized transfer errors on the
certainty equivalents dataset for 17 models across 44 domains.
}
\keyword{datasets}
