% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_interval.R
\name{check_everywhere_dominance}
\alias{check_everywhere_dominance}
\title{Check if method 1 everywhere-upper-dominates method 2
at the tau-th quantile}
\usage{
check_everywhere_dominance(errmat1, errmat2, tau)
}
\arguments{
\item{errmat1}{a square matrix with \code{errmat[i, j]} being the transfer error using the i-th domain for training and j-th domain for testing for method 1}

\item{errmat2}{the error matrix for method 2}

\item{tau}{the quantile at which the methods are compared}
}
\value{
a logical indicating whether method 1 everywhere-upper-dominates method 2 at the tau-th quantile
}
\description{
Check if method 1 everywhere-upper-dominates method 2
at the tau-th quantile
}
