% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ustats_mean_CI.R
\name{ETE_interval}
\alias{ETE_interval}
\title{Two-sided confidence interval for the expectation of
transfer errors}
\usage{
ETE_interval(errmat, alpha, bd = 1)
}
\arguments{
\item{errmat}{a square matrix with \code{errmat[i, j]} being the transfer error using the i-th domain for training and j-th domain for testing}

\item{alpha}{confidence level}

\item{bd}{the upper bound of the loss function}
}
\value{
a 2-d vector giving the left and right ends of the confidence interval
}
\description{
Two-sided confidence interval for the expectation of
transfer errors
}
