% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_interval.R
\name{Gamma_curve}
\alias{Gamma_curve}
\title{The worst-case transfer error \eqn{\bar{e}_{\tau}(\Gamma)} as
a function of \eqn{\tau} for a given \eqn{\Gamma}.}
\usage{
Gamma_curve(errmat, Gamma)
}
\arguments{
\item{errmat}{a square matrix with \code{errmat[i, j]} being the transfer error using the i-th domain for training and j-th domain for testing}

\item{Gamma}{the Gamma value for selection bias}
}
\value{
a data.frame with two columns: \code{prob} for all turning points of \eqn{\tau} where \eqn{\bar{e}_{\tau}(\Gamma)} changes and \code{val} for the corresponding bounds
}
\description{
The worst-case transfer error \eqn{\bar{e}_{\tau}(\Gamma)} as
a function of \eqn{\tau} for a given \eqn{\Gamma}.
}
