% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ustats_mean_CI.R
\name{HBM_mean_upper}
\alias{HBM_mean_upper}
\title{Upper/Lower confidence bound of mean of U-statistics of order k
via Hoeffding-Bentkus-Maurer inequalities}
\usage{
HBM_mean_upper(U, n, k, alpha)
}
\arguments{
\item{U}{the value of U-statistic}

\item{n}{sample size}

\item{k}{order of U-statistic}

\item{alpha}{confidence level}
}
\description{
Upper/Lower confidence bound of mean of U-statistics of order k
via Hoeffding-Bentkus-Maurer inequalities
}
