% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ustats_quantile_CI.R
\name{HBM_quantile_lower}
\alias{HBM_quantile_lower}
\title{Upper/Lower confidence bound for uhat defined in the note
via Hoeffding-Bentkus-Maurer inequalities}
\usage{
HBM_quantile_lower(beta, n, k, alpha)
}
\arguments{
\item{beta}{quantile}

\item{n}{sample size}

\item{k}{order of U-statistic}

\item{alpha}{confidence level}
}
\description{
Upper/Lower confidence bound for uhat defined in the note
via Hoeffding-Bentkus-Maurer inequalities
}
