% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ustats_concentration.R
\name{inv_ustat_HBM_upper}
\alias{inv_ustat_HBM_upper}
\title{Inverse of \eqn{B_{n, k}(x; \mu)} at \eqn{alpha} given \eqn{x} or
\eqn{\mu} val measures "x" if type = "mu" and "mu" if
type = "x". Setting type = "x" yields \eqn{\inf{y: B_{n, k}(y; \mu) \ge \alpha}} and setting type = "mu" yields
\eqn{\inf{\nu: B_{n, k}(x; \nu) \ge \alpha}}}
\usage{
inv_ustat_HBM_upper(val, n, k, alpha, type = c("x", "mu"))
}
\arguments{
\item{val}{the value for either "x" or "mu"}

\item{n}{sample size}

\item{k}{order of the U statistic}

\item{alpha}{confidence level}

\item{type}{see Details}
}
\value{
the inverse of \eqn{B_{n, k}(x; \mu)}
}
\description{
Inverse of \eqn{B_{n, k}(x; \mu)} at \eqn{alpha} given \eqn{x} or
\eqn{\mu} val measures "x" if type = "mu" and "mu" if
type = "x". Setting type = "x" yields \eqn{\inf{y: B_{n, k}(y; \mu) \ge \alpha}} and setting type = "mu" yields
\eqn{\inf{\nu: B_{n, k}(x; \nu) \ge \alpha}}
}
