% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ustats_quantile_CI.R
\name{QTE_interval}
\alias{QTE_interval}
\title{Two-sided confidence interval for the expectation of
transfer errors}
\usage{
QTE_interval(errmat, alpha, betas = seq(0.2, 0.8, 0.01))
}
\arguments{
\item{errmat}{a square matrix with \code{errmat[i, j]} being the transfer error using the i-th domain for training and j-th domain for testing}

\item{alpha}{confidence level}

\item{betas}{a vector of quantiles at which confidence intervals are computed}
}
\value{
a data.frame with three columns: \code{beta} for the quantile, \code{lb} for the lower confidence bound, and \code{ub} for the upper confidence bound.
}
\description{
Two-sided confidence interval for the expectation of
transfer errors
}
